refactor(server): move deletion logic out of cron folder
