+++ packages/server/lib/deletion/deleteEnvironmentData.ts
