+++ packages/server/lib/deletion/deleteSyncConfigData.ts
@@ -3,7 +3,7 @@ import { getFunctionFileLocations, hardDeleteEndpoints, hardDeleteSyncConfig } f
 
 import { batchDelete } from './batchDelete.js';
 import { deleteSyncs } from './deleteSyncs.js';
-import { tasks } from '../../tasks/index.js';
+import { tasks } from '../tasks/index.js';
 
 import type { BatchDeleteSharedOptions } from './batchDelete.js';
 import type { Sync } from '@nangohq/shared';
